Enhancement: Overview page & ExternalLinks updates
