+++ src/components/app/details/AboutAppInfoModal.tsx
@@ -1,38 +1,36 @@
 import React, { useEffect, useState } from 'react'
 import { ReactComponent as Close } from '../../../assets/icons/ic-cross.svg'
-import moment from 'moment'
-import { Moment12HourFormat } from '../../../config'
-import { Progressing, useAsync } from '../../common'
+import { Progressing, showError, VisibleModal } from '../../common'
 import TagLabelSelect from './TagLabelSelect'
 import { ReactComponent as Error } from '../../../assets/icons/ic-warning.svg'
-import { validateTags } from '../appLabelCommon'
-import { getTeamList } from '../../../services/service'
+import { createOption, handleKeyDown, TAG_VALIDATION_MESSAGE, validateTags } from '../appLabelCommon'
 import ReactSelect from 'react-select'
 import InfoColourBar from '../../common/infocolourBar/InfoColourbar'
 import { DropdownIndicator, getCommonSelectStyle, Option } from '../../v2/common/ReactSelect.utils'
-import { NumberOptionType } from '../types'
+import { AboutAppInfoModalProps, LabelTagsType, NumberOptionType } from '../types'
+import { createAppLabels } from '../service'
+import { toast } from 'react-toastify'
 
 export default function AboutAppInfoModal({
-    appId,
-    onClose,
-    appMetaResult,
     isLoading,
-    labelTags,
-    handleInputChange,
-    handleTagsChange,
-    handleKeyDown,
-    handleCreatableBlur,
-    handleSubmit,
-    submitting,
-}) {
-    const [fetchingProjects, projectsListRes] = useAsync(() => getTeamList(), [appId])
+    appId,
+    isChangeProjectView,
+    onClose,
+    appMetaInfo,
+    currentLabelTags,
+    getAppMetaInfoRes,
+    fetchingProjects,
+    projectsList,
+}: AboutAppInfoModalProps) {
     const [projectsOptions, setProjectsOption] = useState<NumberOptionType[]>([])
-    const [selectedProject, setSelectedProject] = useState<NumberOptionType>(null)
+    const [selectedProject, setSelectedProject] = useState<NumberOptionType>()
+    const [submitting, setSubmitting] = useState(false)
+    const [labelTags, setLabelTags] = useState<LabelTagsType>(currentLabelTags)
 
     useEffect(() => {
-        if (!fetchingProjects && projectsListRes?.result && appMetaResult) {
-            const _projectsOption = projectsListRes.result.map((_project) => {
-                if (!selectedProject && _project.name === appMetaResult.projectName) {
+        if (appMetaInfo && !fetchingProjects && Array.isArray(projectsList)) {
+            const _projectsOption = projectsList.map((_project) => {
+                if (!selectedProject && _project.name === appMetaInfo.projectName) {
                     setSelectedProject({ label: _project.name, value: _project.id })
                 }
                 return { label: _project.name, value: _project.id }
@@ -40,17 +38,13 @@ export default function AboutAppInfoModal({
 
             setProjectsOption(_projectsOption)
         }
-    }, [fetchingProjects, projectsListRes, appMetaResult])
-
-    const handleCancelAction = () => {
-        onClose(false)
-    }
+    }, [appMetaInfo, fetchingProjects, projectsList])
 
     const renderAboutModalInfoHeader = () => {
         return (
-            <div className="flex dc__content-space pt-16 pb-16 pl-20 pr-20">
-                <h2 className="fs-20 cn-9 fw-6 m-0">About</h2>
-                <button className="dc__transparent" onClick={handleCancelAction}>
+            <div className="flex dc__content-space pt-16 pb-16 pl-20 pr-20 dc__border-bottom">
+                <h2 className="fs-20 cn-9 fw-6 m-0">{isChangeProjectView ? 'Change project' : 'Manage tags'}</h2>
+                <button className="dc__transparent" onClick={onClose}>
                     <Close className="icon-dim-20 cursor" />
                 </button>
             </div>
@@ -68,7 +62,7 @@ export default function AboutAppInfoModal({
         }
     }
 
-    const handleProjectSelection = (selected) => {
+    const handleProjectSelection = (selected: NumberOptionType) => {
         setSelectedProject(selected)
     }
 
@@ -98,60 +92,157 @@ export default function AboutAppInfoModal({
         )
     }
 
-    const handleSaveAction = (e) => {
+    const setAppTagLabel = () => {
+        const newTag = labelTags.inputTagValue.split(',').map((e) => {
+            e = e.trim()
+            return createOption(e)
+        })
+        setLabelTags({
+            inputTagValue: '',
+            tags: [...labelTags.tags, ...newTag],
+            tagError: '',
+        })
+    }
+
+    function validateForm(): boolean {
+        if (
+            labelTags.tags.length !== labelTags.tags.map((tag) => tag.value).filter((tag) => validateTags(tag)).length
+        ) {
+            setLabelTags((labelTags) => ({ ...labelTags, tagError: TAG_VALIDATION_MESSAGE.error }))
+            return false
+        }
+        return true
+    }
+
+    const handleKeyDownEvent = (e) => {
+        handleKeyDown(labelTags, setAppTagLabel, e)
+    }
+
+    function handleInputChange(inputTagValue) {
+        setLabelTags((tags) => ({ ...tags, inputTagValue: inputTagValue, tagError: '' }))
+    }
+
+    function handleTagsChange(newValue: any, actionMeta: any) {
+        setLabelTags((tags) => ({ ...tags, tags: newValue || [], tagError: '' }))
+    }
+
+    function handleCreatableBlur(e) {
+        labelTags.inputTagValue = labelTags.inputTagValue.trim()
+        if (!labelTags.inputTagValue) {
+            return
+        } else {
+            setLabelTags({
+                inputTagValue: '',
+                tags: [...labelTags.tags, createOption(e.target.value)],
+                tagError: '',
+            })
+        }
+    }
+
+    const handleSaveAction = async (e) => {
         e.preventDefault()
-        handleSubmit(e, selectedProject)
+        const validForm = validateForm()
+        if (!validForm) {
+            return
+        }
+        setSubmitting(true)
+
+        const _optionTypes = []
+        if (labelTags.tags && labelTags.tags.length > 0) {
+            labelTags.tags.forEach((_label) => {
+                let colonIndex = _label.value.indexOf(':')
+                let splittedTagBeforeColon = _label.value.substring(0, colonIndex)
+                let splittedTagAfterColon = _label.value.substring(colonIndex + 1)
+                _optionTypes.push({
+                    key: splittedTagBeforeColon,
+                    value: splittedTagAfterColon,
+                })
+            })
+        }
+
+        const payload = {
+            id: parseInt(appId),
+            labels: _optionTypes,
+            teamId: selectedProject.value,
+        }
+
+        try {
+            await createAppLabels(payload)
+            if (appMetaInfo.projectName === selectedProject.label) {
+                toast.success('Successfully saved')
+            } else {
+                toast.success(`Application '${appMetaInfo.appName}' is moved to project '${selectedProject.label}'`)
+            }
+
+            // Fetch the latest project & labels details
+            await getAppMetaInfoRes()
+        } catch (err) {
+            if (err['code'] === 403 && appMetaInfo.projectName !== selectedProject.label) {
+                toast.error(`You don't have the required access to the target project ${selectedProject.label}`)
+            } else {
+                showError(err)
+            }
+        } finally {
+            onClose()
+            setSubmitting(false)
+        }
+    }
+
+    const projectChangeMessage = () => {
+        return (
+            <>
+                <span className="fs-13 fw-4 lh-20 cn-9">Project change may lead to:</span>
+                <ol className="fs-13 fw-4 lh-20 cn-9 pl-20 pr-4">
+                    <li>Current users losing access to this application.</li>
+                    <li>
+                        Users getting an access to the application automatically, if they have an access to the selected
+                        project.
+                    </li>
+                </ol>
+            </>
+        )
     }
 
     const renderAboutModalInfo = () => {
         return (
             <>
                 <div className="cn-7 p-20">
-                    <div>
-                        <div className="fs-12 fw-4 lh-20">App name</div>
-                        <div className="fs-14 fw-6 lh-20 mb-20">{appMetaResult?.appName}</div>
-                    </div>
-                    <div>
-                        <div className="fs-12 fw-4 lh-20">Created on</div>
-                        <div className="fs-14 fw-6 lh-20 mb-20">
-                            {moment(appMetaResult?.createdOn).format(Moment12HourFormat)}
-                        </div>
-                    </div>
-                    <div>
-                        <div className="fs-12 fw-4 lh-20">Created by</div>
-                        <div className="fs-14 fw-6 lh-20 mb-20">{appMetaResult?.createdBy}</div>
-                    </div>
-                    <div className="mb-20">
-                        <div className="fs-12 fw-4 lh-20 mb-2">Project</div>
-                        {renderProjectSelect()}
-                        {selectedProject && appMetaResult && selectedProject.label !== appMetaResult.projectName && (
-                            <InfoColourBar
-                                classname="warn cn-9 lh-20"
-                                Icon={Error}
-                                message="Changing the project may result in current users losing access to this application. Users with access to the target project will gain access after the application is moved."
-                                iconClass="warning-icon"
-                                styles={{
-                                    padding: '8px 12px',
-                                }}
+                    {isChangeProjectView ? (
+                        <>
+                            <div className="fs-12 fw-4 lh-20 mb-2">Project</div>
+                            {renderProjectSelect()}
+                            {selectedProject && appMetaInfo && selectedProject.label !== appMetaInfo.projectName && (
+                                <InfoColourBar
+                                    classname="warn cn-9 lh-20"
+                                    Icon={Error}
+                                    message={projectChangeMessage()}
+                                    iconClass="warning-icon"
+                                    styles={{
+                                        padding: '8px 12px',
+                                    }}
+                                />
+                            )}
+                        </>
+                    ) : (
+                        <>
+                            <TagLabelSelect
+                                validateTags={validateTags}
+                                labelTags={labelTags}
+                                onInputChange={handleInputChange}
+                                onTagsChange={handleTagsChange}
+                                onKeyDown={handleKeyDownEvent}
+                                onCreatableBlur={handleCreatableBlur}
                             />
-                        )}
-                    </div>
-                    <TagLabelSelect
-                        validateTags={validateTags}
-                        labelTags={labelTags}
-                        onInputChange={handleInputChange}
-                        onTagsChange={handleTagsChange}
-                        onKeyDown={handleKeyDown}
-                        onCreatableBlur={handleCreatableBlur}
-                    />
-                    {renderValidationMessaging()}
+                            {renderValidationMessaging()}
+                        </>
+                    )}
                 </div>
                 <div className="form__buttons dc__border-top pt-16 pb-16 pl-20 pr-20">
                     <button
                         className="cta cancel flex h-36 mr-12"
                         type="button"
                         disabled={submitting}
-                        onClick={handleCancelAction}
+                        onClick={onClose}
                         tabIndex={6}
                     >
                         Cancel
@@ -171,15 +262,17 @@ export default function AboutAppInfoModal({
     }
 
     return (
-        <div>
-            {renderAboutModalInfoHeader()}
-            {isLoading || fetchingProjects ? (
-                <div className="flex" style={{ minHeight: '400px' }}>
-                    <Progressing pageLoader />
-                </div>
-            ) : (
-                renderAboutModalInfo()
-            )}
-        </div>
+        <VisibleModal className="app-status__material-modal">
+            <div className="modal__body br-8 bcn-0 mt-0-imp p-0 dc__no-top-radius">
+                {renderAboutModalInfoHeader()}
+                {isLoading || fetchingProjects ? (
+                    <div className="flex" style={{ minHeight: '400px' }}>
+                        <Progressing pageLoader />
+                    </div>
+                ) : (
+                    renderAboutModalInfo()
+                )}
+            </div>
+        </VisibleModal>
     )
 }
