+++ src/components/app/details/appOverview/AppOverview.scss
@@ -1,0 +1,27 @@
+.app-overview-container {
+    grid-template-columns: 250px 1fr;
+
+    .app-overview-wrapper {
+        .app-level__external-links {
+            padding: 0;
+            margin-bottom: 0 !important;
+            min-height: 0;
+        }
+
+        .env-deployments-info-wrapper {
+
+            .env-deployments-info-header,
+            .env-deployments-info-row {
+                grid-template-columns: 250px 1fr;
+                column-gap: 16px;
+                grid-auto-rows: 36px;
+            }
+        }
+    }
+}
+
+.app-status__material-modal {
+    .modal__body {
+        width: 600px;
+    }
+}
