+++ src/components/externalLinks/ExternalLinks.type.ts
@@ -21,6 +21,7 @@ export interface ExternalLink {
     id?: number
     monitoringToolId: number
     name: string
+    description: string
     url: string
     clusterIds: any[]
     updatedOn?: string
@@ -31,6 +32,7 @@ export interface LinkAction {
     invalidTool?: boolean
     name: string
     invalidName?: boolean
+    description: string
     clusters: MultiValue<OptionType>
     invalidClusters?: boolean
     urlTemplate: string
@@ -101,6 +103,7 @@ export interface AppLevelExternalLinksType {
     helmAppDetails?: HelmAppDetails
     externalLinks: ExternalLink[]
     monitoringTools: OptionTypeWithIcon[]
+    isOverviewPage?: boolean
 }
 
 export interface NodeLevelExternalLinksType  {
@@ -113,6 +116,7 @@ export interface NodeLevelExternalLinksType  {
 }
 
 export interface ExternalLinksAndToolsType {
+    fetchingExternalLinks?: boolean
     externalLinks: ExternalLink[]
     monitoringTools: OptionTypeWithIcon[]
 }
