+++ src/components/externalLinks/externalLinks.component.scss
@@ -1,230 +1,3 @@
-.add-external-link-dialog {
-    .modal__body {
-        width: calc(100% - 80px);
-        max-width: 1000px;
-        height: calc(100% - 40px);
-        border-radius: 0 0 8px 8px;
-        margin-top: 0;
-        padding: 16px 0;
-
-        .modal__divider {
-            border: 0;
-            height: 1px;
-            box-shadow: inset 0 -1px 0 0 var(--N200);
-        }
-
-        .modal__dropdown-divider {
-            height: 3px;
-            width: 100%;
-            margin-left: -40px;
-            position: absolute;
-            top: 59px;
-            border-bottom: 1px solid var(--N200);
-        }
-
-        .modal__header {
-            padding: 0 20px 16px;
-            margin: 0;
-
-            h1 {
-                font-size: 16px;
-            }
-        }
-
-        .modal__content {
-            display: flex;
-            height: calc(100% - 112px);
-
-            .configure-link-action-container,
-            .configure-link-info-container {
-                overflow-wrap: break-word;
-                overflow-y: scroll;
-            }
-
-            .configure-link-action-container {
-                width: 720px;
-                padding: 20px;
-                border-right: 1px solid var(--N200);
-
-                .link-add-another {
-                    display: flex;
-                    align-items: center;
-                    color: var(--B500);
-
-                    svg path {
-                        fill: var(--B500);
-                    }
-                }
-
-                .configure-link-action-wrapper,
-                label {
-                    color: var(--N700);
-                    font-size: 13px;
-                    font-weight: 400;
-                    margin-bottom: 6px;
-                }
-
-                .configure-link-action-wrapper {
-                    .error-label {
-                        .error-label-icon {
-                            width: 20px;
-                            height: 20px;
-
-                            svg g path:nth-of-type(2) {
-                                fill: var(--R500);
-                            }
-                        }
-                    }
-
-                    .link-monitoring-tool {
-                        display: flex;
-                        align-items: flex-start;
-
-                        .link-name {
-                            display: flex;
-                            flex-direction: column;
-                            width: 200px;
-
-                            input {
-                                width: 200px;
-                                height: 32px;
-                                padding: 10px 6px;
-                                background-color: var(--N50);
-                                color: var(--N900);
-                                border: 1px solid var(--N200);
-                                border-radius: 4px;
-                            }
-                        }
-
-                        .monitoring-tool {
-                            width: 150px;
-
-                            .absolute-option {
-                                position: absolute;
-                                padding: 0 10px;
-                            }
-                        }
-
-                        .link-clusters {
-                            width: 278px;
-                        }
-
-                        .link-delete {
-                            width: 20px;
-                            height: 20px;
-                        }
-                    }
-                }
-
-                .link-text-area {
-                    display: flex;
-                    flex-direction: column;
-
-                    textarea {
-                        width: 648px;
-                        min-height: 72px;
-                        resize: vertical;
-                        padding: 10px 8px;
-                        background-color: var(--N50);
-                        color: var(--N900);
-                        border: 1px solid var(--N200);
-                        border-radius: 4px;
-
-                        &::placeholder {
-                            color: var(--N500);
-                        }
-
-                        &:-ms-input-placeholder {
-                            color: var(--N500);
-                        }
-                    }
-                }
-            }
-
-            .configure-link-info-container {
-                width: 279px;
-                padding: 20px;
-
-                .configure-link-info-heading {
-                    display: flex;
-                    align-items: center;
-                    font-size: 13px;
-                    font-weight: 600;
-                    line-height: 1.5;
-
-                    svg {
-                        width: 20px;
-                        height: 20px;
-                        margin-right: 10px;
-
-                        path {
-                            fill: var(--V500);
-                        }
-                    }
-                }
-
-                .configure-link-info-list {
-                    padding: 20px 15px 0;
-                    margin-bottom: 0;
-                    color: var(--N900);
-
-                    li,
-                    p {
-                        font-size: 13px;
-                    }
-
-                    li,
-                    p,
-                    a {
-                        line-height: 1.5;
-                    }
-
-                    li {
-                        font-weight: 600;
-                        margin-bottom: 4px;
-                    }
-
-
-                    ul {
-                        list-style: none;
-                        padding: 0;
-                        margin: 0 0 12px;
-
-                        li {
-                            font-size: 12px;
-                            font-weight: 400;
-                            background-color: var(--N100);
-                            width: fit-content;
-                            padding: 2px 4px;
-                            margin-bottom: 8px;
-                        }
-                    }
-
-                    p {
-                        font-weight: 400;
-
-                        &:last-of-type {
-                            font-style: italic;
-                        }
-                    }
-
-                    a {
-                        color: var(--B500);
-                        font-size: 12px;
-                        font-weight: 600;
-                    }
-                }
-            }
-        }
-
-        .modal__buttons {
-            display: flex;
-            justify-content: flex-end;
-            padding: 16px 20px;
-        }
-    }
-}
-
 .app-level__external-links {
     width: 100%;
     min-height: 56px;
