+++ src/components/externalLinks/externalLinks.scss
@@ -1,6 +1,8 @@
 .external-links-container {
-    min-width: 720px;
-    padding: 16px 8%  20px;
+    min-width: 962px;
+    width: 90%;
+    margin: auto;
+    padding: 20px 24px;
 
     &.error-view {
         min-width: 100%;
@@ -51,11 +53,9 @@
     }
 
     .external-links-wrapper {
-        max-width: 962px;
+        max-width: 1200px;
         margin: auto;
         .cta-search-filter-container {
-            margin-top: 20px;
-
             .add-link.cta {
                 height: 32px;
             }
@@ -160,7 +160,7 @@
 
         .external-link {
             display: grid;
-            grid-template-columns: 24px 144px 84px minmax(100px, 59%) 56px;
+            grid-template-columns: 24px 144px minmax(100px, calc(59% - 310px)) 84px minmax(100px, 59%) 56px;
             column-gap: 16px;
             padding: 12px 20px;
 
@@ -188,7 +188,7 @@
     .external-links__header {
         height: 40px;
         display: grid;
-        grid-template-columns: 24px 144px 84px minmax(100px, 59%) 56px;
+        grid-template-columns: 24px 144px minmax(100px, calc(59% - 322px)) 84px minmax(100px, 59%) 56px;
         column-gap: 16px;
         align-items: center;
         justify-content: stretch;
