+++ src/config/routes.ts
@@ -23,6 +23,7 @@ export const URLS = {
     APP_DETAILS_LOG: 'log-analyzer', // for V2
     APP_DETAILS_DEFAULT: 'default-view',
     APP_TRIGGER: 'trigger',
+    APP_OVERVIEW: 'overview',
     APP_CI_DETAILS: 'ci-details',
     APP_CD_DETAILS: 'cd-details',
     APP_DEPLOYMENT_METRICS: 'deployment-metrics',
