+++ src/css/formulae.scss
@@ -62,6 +62,10 @@
     align-items: center;
 }
 
+.display-grid {
+    display: grid;
+}
+
 // margins and paddings from 0-40
 $margins: (mb: margin-bottom, mt: margin-top, ml: margin-left, mr: margin-right, m: margin, pb: padding-bottom, pt: padding-top, pl: padding-left, pr: padding-right, p: padding);
