+++ src/services/service.types.ts
@@ -89,6 +89,7 @@ export interface AppEnvironment {
     infraMetrics: boolean;
     prod: boolean;
     chartRefId?: number
+    lastDeployedTime?: string
 }
 
 export interface AppOtherEnvironment extends ResponseType {
